+++ pkg/daemon/ceph/mon/daemon_test.go
@@ -18,13 +18,14 @@ package mon
 import (
 	"testing"
 
+	cephconfig "github.com/rook/rook/pkg/daemon/ceph/config"
 	"github.com/stretchr/testify/assert"
 )
 
 func TestMonFlattening(t *testing.T) {
 
 	// single endpoint
-	mons := map[string]*CephMonitorConfig{
+	mons := map[string]*cephconfig.MonInfo{
 		"foo": {Name: "foo", Endpoint: "1.2.3.4:5000"},
 	}
 	flattened := FlattenMonEndpoints(mons)
@@ -35,7 +36,7 @@ func TestMonFlattening(t *testing.T) {
 	assert.Equal(t, "1.2.3.4:5000", parsed["foo"].Endpoint)
 
 	// multiple endpoints
-	mons["bar"] = &CephMonitorConfig{Name: "bar", Endpoint: "2.3.4.5:6000"}
+	mons["bar"] = &cephconfig.MonInfo{Name: "bar", Endpoint: "2.3.4.5:6000"}
 	flattened = FlattenMonEndpoints(mons)
 	parsed = ParseMonEndpoints(flattened)
 	assert.Equal(t, 2, len(parsed))
