+++ pkg/operator/test/info.go
@@ -20,22 +20,22 @@ package test
 import (
 	"fmt"
 
-	"github.com/rook/rook/pkg/daemon/ceph/mon"
+	cephconfig "github.com/rook/rook/pkg/daemon/ceph/config"
 )
 
 // CreateConfigDir creates a test cluster
-func CreateConfigDir(monCount int) *mon.ClusterInfo {
-	c := &mon.ClusterInfo{
+func CreateConfigDir(monCount int) *cephconfig.ClusterInfo {
+	c := &cephconfig.ClusterInfo{
 		FSID:          "12345",
 		Name:          "default",
 		MonitorSecret: "monsecret",
 		AdminSecret:   "adminsecret",
-		Monitors:      map[string]*mon.CephMonitorConfig{},
+		Monitors:      map[string]*cephconfig.MonInfo{},
 	}
 	mons := []string{"a", "b", "c", "d", "e"}
 	for i := 0; i < monCount; i++ {
 		id := mons[i]
-		c.Monitors[id] = &mon.CephMonitorConfig{
+		c.Monitors[id] = &cephconfig.MonInfo{
 			Name:     id,
 			Endpoint: fmt.Sprintf("1.2.3.%d:6790", i+1),
 		}
